Refactor client into modular package
