+++ main.py
@@ -1,6 +1,6 @@
 """Entry point for launching the VatiVision Pro client application."""
 
-from vativision_pro_client import main as run_app
+from vativision_pro.ui.main_window import main as run_app
 
 
 if __name__ == "__main__":
