+++ vativision_pro/media/cursor.png
